% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_singscore.R
\name{compute_singscore}
\alias{compute_singscore}
\title{Singscore}
\usage{
compute_singscore(expr_data, signatures)
}
\arguments{
\item{expr_data}{A gene expression matrix (or data frame) with gene symbols as
row names and samples as columns.}

\item{signatures}{A list of named character vectors corresponding to gene signatures.}
}
\value{
A tibble with the first column indicating sample identifiers (\code{sample_id})
and all the other columns ....
}
\description{
Obtain single sample GSEA scores for a list of gene signatures,
as implemented in \emph{Foroutan et al., 2018}.
}
\examples{
set.seed(123)

test_expr <- data.frame(
    sample1 = rnorm(10, mean = 8, sd = 2),
    sample2 = rnorm(10, mean = 7, sd = 2),
    sample3 = rnorm(10, mean = 8, sd = 3),
    row.names = paste0("gene", 1:10)
)

test_list <- list(
    gs1 = paste0("gene", c(1, 2, 6)),
    gs2 = paste0("gene", 1:5),
    gs3 = paste0("gene", c(3, 1, 9)),
    gs4 = paste0("gene", c(1, 3, 9))
)

compute_singscore(test_expr, test_list)

}
\references{
Foroutan, M., Bhuva, D. D., Lyu, R., Horan, K., Cursons, J., &
Davis, M. J. (2018). Single sample scoring of molecular phenotypes.
\emph{BMC bioinformatics}, 19(1), 404.
\href{https://doi.org/10.1186/s12859-018-2435-4}{doi: 10.1186/s12859-018-2435-4}.

Hänzelmann, S., Castelo, R., & Guinney, J. (2013). GSVA: gene set variation
analysis for microarray and RNA-seq data. \emph{BMC bioinformatics}, 14, 7.
\href{https://doi.org/10.1186/1471-2105-14-7}{doi: 10.1186/1471-2105-14-7}.
}
\seealso{
\code{\link[singscore]{simpleScore}}
}
