% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ssgsea.R
\name{compute_ssgsea}
\alias{compute_ssgsea}
\title{Single Sample GSEA}
\usage{
compute_ssgsea()
}
\arguments{
\item{expr_data}{A gene expression matrix (or data frame) with gene symbols as
row names and samples as columns.}

\item{signature}{A list of character vectors corresponding to gene signatures.}
}
\value{
A data frame with
}
\description{
Obtain single sample GSEA (ssGSEA) scores, as implemented in XXX, from a cohort of samples.
}
\examples{
m <- matrix()
}
