% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig_keywords.R
\name{get_sig_keywords}
\alias{get_sig_keywords}
\title{Get keywords associated to signatures}
\usage{
get_sig_keywords()
}
\value{
A character vector with valid keywords.
}
\description{
Obtain valid keywords for the argument \code{signatures} of \code{hack_sig()} and
\code{check_sig()} in order to filter for specific gene signatures.
Some of the returned strings are redundant, such as \code{"interferon gamma"} and \code{"ifng"}.
}
\examples{
get_sig_keywords()
}
\seealso{
\code{\link[=check_sig]{check_sig()}}, \code{\link[=hack_sig]{hack_sig()}}
}
