% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack_class.R
\name{hack_class}
\alias{hack_class}
\title{Hack signature classes}
\usage{
hack_class(sig_data, cutoff = "original")
}
\arguments{
\item{sig_data}{A tibble result of a call to \code{\link[=hack_sig]{hack_sig()}}.}

\item{cutoff}{A character specifying which function to use to categorize
samples by signature scores. Can be one of:\preformatted{* `"original"` (default), apply the original publication method;
* `"mean"/"median"`, samples will be classified as `"low"` or `"high"` with respect
  to the mean/median signature score value, respectively;
* `"tertiles"`, samples will be classified as `"<= T1"` (score lower than
  first tertile), `"(T1, T2]"` (score between first and second tertiles),
  `"> T2"` (score higher than second tertile);
* `"quartiles"`, samples will be classified as `"<= Q1"` (score lower than
  first quartile), `"(Q1, Q2]"` (score between first and second quartiles),
  `"(Q2, Q3]"` (score between second and third quartiles), `"> Q3"` (score
  higher than third quartile).
}}
}
\value{
A tibble with the same dimension as \code{sig_data}, a column \code{sample_id}
indicating sample identifiers and one column for each input signature giving
signature classes.
}
\description{
\code{hack_class()} is supposed to be used in combination after \code{\link[=hack_sig]{hack_sig()}} in
order to classify your samples in one of two or more signature classes.
}
\examples{
library(dplyr)
hack_sig(test_expr, "immune") \%>\% hack_class()

}
\seealso{
\code{\link[=hack_sig]{hack_sig()}}
}
