% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack_ifng.R
\name{hack_ifng}
\alias{hack_ifng}
\title{Hack Interferon Gamma Signatures}
\usage{
hack_ifng(expr_data, signature = "all", method = "default")
}
\arguments{
\item{signature}{character}

\item{method}{character}
}
\value{
dataframe
}
\description{
Hack Interferon Gamma Signatures
}
