% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack_immunophenoscore.R
\name{hack_immunophenoscore}
\alias{hack_immunophenoscore}
\title{Immunophenoscore}
\usage{
hack_immunophenoscore(expr_data)
}
\arguments{
\item{expr_data}{A gene expression matrix (or data frame) with gene symbols as
row names and samples as columns.}
}
\value{
A data frame with
}
\description{
Obtain various immune biomarkers scores, which combined together give the
immunophenoscore (\emph{Charoentong et al., 2017}).
}
\examples{

}
\references{
Charoentong, P., Finotello, F., Angelova, M., Mayer, C.,
Efremova, M., Rieder, D., Hackl, H., & Trajanoski, Z. (2017). Pan-cancer
Immunogenomic Analyses Reveal Genotype-Immunophenotype Relationships and
Predictors of Response to Checkpoint Blockade. \emph{Cell reports}, 18(1), 248–262.
\href{https://doi.org/10.1016/j.celrep.2016.12.019}{doi: 10.1016/j.celrep.2016.12.019}.
}
