% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack_sig.R
\name{hack_sig}
\alias{hack_sig}
\title{Score samples by gene signatures}
\usage{
hack_sig(expr_data, signatures = "all", method = "default", ...)
}
\arguments{
\item{expr_data}{A gene expression matrix (or data frame) with gene symbols as
row names and samples as columns.}

\item{signatures}{It can be a list of signatures or a character string indicating
a keyword for a group of signatures (e.g. "immune" or "ifng" for immune and
interferon gamma signatures respectively). The default (\code{all}) will cause the
function to compute single sample scores for all the signatures implemented
in \code{hacksig}.}

\item{method}{A character string specifying which method to use for computing
the single sample score for each signature. You can choose one of:
\itemize{
\item \code{default}, the original method used by the authors of the signature;
\item \code{zscore}, the z-score method implemented in \code{compute_zscore()};
\item \code{ssgsea}, the single sample GSEA method implemented in \code{compute_ssgsea()};
\item \code{singscore}, the singscore method implemented in \code{compute_singscore()};
}}

\item{...}{Additional arguments passed to \code{compute_zscore()}, \code{compute_ssgsea()}
or \code{compute_singscore()}, depending on the choice of \code{method}.}

\item{direction}{A character specifying the score computation depending on the
direction of the signatures. Can be on of:
\itemize{
\item \code{none}, undirected signatures, that is you don't know whether the genes are
up- or down-regulated (default);
\item \code{up}, all genes in the signature are supposed to be up-regulated;
\item \code{down}, all genes in the signature are supposed to be down-regulated;
\item \code{both}, a signature is composed of up- and down sub-signatures.
You must supply it as a nested list.
}}

\item{norm}{A character string specifying the type of normalization affecting
the single sample GSEA scores. Can be one of:
\itemize{
\item \code{raw}, obtain raw scores (\emph{default});
\item \code{separate}, normalize raw scores across samples for each signature separately.
\item \code{all}, normalize raw scores both across samples and signatures.
}}

\item{alpha}{A numeric scalar. Exponent in the running sum of the single sample
GSEA score calculation which weights the gene ranks. Defaults to \eqn{\alpha = 0.25}.}
}
\value{
A tibble with the first column indicating sample identifiers (\code{sample_id})
and all the other columns ....
}
\description{
\code{hack_sig()} is the main function of the package, which computes single sample
scores in one of different ways. You can choose to apply the default method
for a signature or you can choose one of three single sample scoring methods:
\itemize{
\item \strong{z-score} (\emph{Lee et al., 2008});
\item \strong{ssGSEA} (\emph{Barbie et al., 2009});
\item \strong{singscore} (\emph{Foroutan et al., 2018}).
}
}
\examples{
set.seed(123)

test_expr <- data.frame(
    sample1 = rnorm(10, mean = 8, sd = 2),
    sample2 = rnorm(10, mean = 7, sd = 2),
    sample3 = rnorm(10, mean = 8, sd = 3),
    row.names = paste0("gene", 1:10)
)

test_list <- list(
    gs1 = paste0("gene", c(1, 2, 6)),
    gs2 = paste0("gene", 1:5),
    gs3 = paste0("gene", c(3, 1, 9)),
    gs4 = paste0("gene", c(1, 3, 9))
)

hack_sig(test_expr, test_list)
}
\references{
Barbie, D. A., Tamayo, P., Boehm, J. S., Kim, S. Y., Moody, S. E., Dunn, I. F.,
Schinzel, A. C., Sandy, P., Meylan, E., Scholl, C., Fröhling, S., Chan, E. M.,
Sos, M. L., Michel, K., Mermel, C., Silver, S. J., Weir, B. A., Reiling, J. H.,
Sheng, Q., Gupta, P. B., … Hahn, W. C. (2009). Systematic RNA interference
reveals that oncogenic KRAS-driven cancers require TBK1. \emph{Nature}, 462(7269),
108–112. \href{https://doi.org/10.1038/nature08460}{doi: 10.1038/nature08460}.

Foroutan, M., Bhuva, D. D., Lyu, R., Horan, K., Cursons, J., &
Davis, M. J. (2018). Single sample scoring of molecular phenotypes.
\emph{BMC bioinformatics}, 19(1), 404.
\href{https://doi.org/10.1186/s12859-018-2435-4}{doi: 10.1186/s12859-018-2435-4}.

Hänzelmann, S., Castelo, R., & Guinney, J. (2013). GSVA: gene set variation
analysis for microarray and RNA-seq data. \emph{BMC bioinformatics}, 14, 7.
\href{https://doi.org/10.1186/1471-2105-14-7}{doi: 10.1186/1471-2105-14-7}.

Lee, E., Chuang, H. Y., Kim, J. W., Ideker, T., & Lee, D. (2008). Inferring
pathway activity toward precise disease classification.
\emph{PLoS computational biology}, 4(11), e1000217.
\href{https://doi.org/10.1371/journal.pcbi.1000217}{doi: 10.1371/journal.pcbi.1000217}.

Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B. L.,
Gillette, M. A., Paulovich, A., Pomeroy, S. L., Golub, T. R., Lander, E. S.,
& Mesirov, J. P. (2005). Gene set enrichment analysis: a knowledge-based
approach for interpreting genome-wide expression profiles.
\emph{Proceedings of the National Academy of Sciences of the United States of America},
102(43), 15545–15550. \href{https://doi.org/10.1073/pnas.0506580102}{doi: 10.1073/pnas.0506580102}.
}
