% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signatures_data}
\alias{signatures_data}
\title{Gene-level information for each signature implemented}
\format{
A tibble with six columns:
\itemize{
\item \code{signature_id}: a unique ID given to the signature;
\item \code{signature_keyword}: a list of keywords which can be used to filter signatures;
\item \code{gene_symbol}: HUGO Gene Symbol ID;
\item \code{gene_entrez_id}: Entrez gene ID;
\item \code{gene_weight}: gene weights used in the signature model;
\item \code{signature_method}: default method implemented in the original publication.
}
}
\usage{
signatures_data
}
\description{
A data set used by the \code{hacksig} functions to retrieve gene-level information
about XXX implemented signatures.
}
\details{
Identifiers in \code{signature_id} correspond to the following DOIs:
\itemize{
\item \code{"rooney2015_cytoact"}, \href{https://doi.org/10.1016/j.cell.2014.12.033}{10.1016/j.cell.2014.12.033}
\item \code{"ayers2017_immexp"}, \href{https://doi.org/10.1172/JCI91190}{10.1172/JCI91190}
\item \code{"she2020_irgs"}, \href{https://doi.org/10.1186/s12935-020-1104-7}{10.1186/s12935-020-1104-7}
\item \code{"wu2020_metabolic"}, \href{https://doi.org/10.1155/2020/6716908}{10.1155/2020/6716908}
\item \code{"fang2020_irgs"}, \href{https://doi.org/10.18632/aging.202842}{10.18632/aging.202842}
}
}
\seealso{
\code{\link[=hack_sig]{hack_sig()}}, \code{\link[=check_sig]{check_sig()}}
}
\keyword{datasets}
