% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_methods.R
\name{ss_methods}
\alias{ss_methods}
\alias{compute_zscore}
\alias{compute_ssgsea}
\alias{compute_singscore}
\title{Single sample methods}
\usage{
compute_zscore(expr_data, signatures)

compute_ssgsea(expr_data, signatures, norm = "raw", alpha = 0.25)

compute_singscore(expr_data, signatures, direction = "none")
}
\arguments{
\item{expr_data}{A gene expression matrix (or data frame) with gene symbols as
row names and samples as columns.}

\item{signatures}{It can be a list of signatures or a character string indicating
a keyword for a group of signatures (e.g. "immune" or "ifng" for immune and
interferon gamma signatures respectively). The default (\code{all}) will cause the
function to compute single sample scores for all the signatures implemented
in \code{hacksig}.}

\item{norm}{A character string specifying the type of normalization affecting
the \strong{single sample GSEA} scores. Can be one of:
\itemize{
\item \code{raw}, obtain raw scores (\emph{default});
\item \code{separate}, normalize raw scores across samples for each signature separately.
\item \code{all}, normalize raw scores both across samples and signatures.
}}

\item{alpha}{A numeric scalar. Exponent in the running sum of the \strong{single sample GSEA}
score calculation which weights the gene ranks. Defaults to \eqn{\alpha = 0.25}.}

\item{direction}{A character specifying the \strong{singscore} computation depending on the
direction of the signatures. Can be on of:
\itemize{
\item \code{none}, undirected signatures, that is you don't know whether the genes are
up- or down-regulated (default);
\item \code{up}, all genes in the signature are supposed to be up-regulated;
\item \code{down}, all genes in the signature are supposed to be down-regulated;
\item \code{both}, a signature is composed of up- and down sub-signatures.
You must supply it as a nested list.
}}
}
\value{
A tibble with one row for each sample in \code{expr_data}, a column \code{sample_id}
indicating sample identifiers and one column for each input signature giving
single sample scores.
}
\description{
These are internal functions to compute single sample scores in three different
ways: \strong{z-score} (\emph{Lee et al., 2008}), \strong{ssGSEA} (\emph{Barbie et al., 2009})
or \strong{singscore} (\emph{Foroutan et al., 2018}).
\code{compute_ssgsea} is called by \code{\link[=hack_estimate]{hack_estimate()}} whereas all the three
methods are called by \code{\link{hack_sig}}.
}
\details{

}
\references{
Barbie, D. A., Tamayo, P., Boehm, J. S., Kim, S. Y., Moody, S. E., Dunn, I. F.,
Schinzel, A. C., Sandy, P., Meylan, E., Scholl, C., Fröhling, S., Chan, E. M.,
Sos, M. L., Michel, K., Mermel, C., Silver, S. J., Weir, B. A., Reiling, J. H.,
Sheng, Q., Gupta, P. B., … Hahn, W. C. (2009). Systematic RNA interference
reveals that oncogenic KRAS-driven cancers require TBK1. \emph{Nature}, 462(7269),
108–112. \href{https://doi.org/10.1038/nature08460}{doi: 10.1038/nature08460}.

Foroutan, M., Bhuva, D. D., Lyu, R., Horan, K., Cursons, J., &
Davis, M. J. (2018). Single sample scoring of molecular phenotypes.
\emph{BMC bioinformatics}, 19(1), 404.
\href{https://doi.org/10.1186/s12859-018-2435-4}{doi: 10.1186/s12859-018-2435-4}.

Hänzelmann, S., Castelo, R., & Guinney, J. (2013). GSVA: gene set variation
analysis for microarray and RNA-seq data. \emph{BMC bioinformatics}, 14, 7.
\href{https://doi.org/10.1186/1471-2105-14-7}{doi: 10.1186/1471-2105-14-7}.

Lee, E., Chuang, H. Y., Kim, J. W., Ideker, T., & Lee, D. (2008). Inferring
pathway activity toward precise disease classification.
\emph{PLoS computational biology}, 4(11), e1000217.
\href{https://doi.org/10.1371/journal.pcbi.1000217}{doi: 10.1371/journal.pcbi.1000217}.

Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B. L.,
Gillette, M. A., Paulovich, A., Pomeroy, S. L., Golub, T. R., Lander, E. S.,
& Mesirov, J. P. (2005). Gene set enrichment analysis: a knowledge-based
approach for interpreting genome-wide expression profiles.
\emph{Proceedings of the National Academy of Sciences of the United States of America},
102(43), 15545–15550. \href{https://doi.org/10.1073/pnas.0506580102}{doi: 10.1073/pnas.0506580102}.
}
\seealso{
\code{\link[=hack_sig]{hack_sig()}}, \code{\link[=hack_estimate]{hack_estimate()}}
}
\keyword{internal}
