% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_sig.R
\name{stratify_sig}
\alias{stratify_sig}
\title{Stratify samples into classes}
\usage{
stratify_sig(sig_data, cutoff = "original", probs = seq(0, 1, 0.25))
}
\arguments{
\item{sig_data}{A tibble result of a call to \code{\link[=hack_sig]{hack_sig()}}.}

\item{cutoff}{A character specifying which function to use to categorize
samples by signature scores. Can be one of:
\itemize{
\item \code{"original"} (default), apply the original publication method; if
categorization is not expected, the median score is used as a threshold;
\item \code{"mean"}/\code{"median"}, samples will be classified as \code{"low"} or \code{"high"} with
respect to the mean/median signature score, respectively;
\item \code{"quantile"}, samples will be classified into signature score quantiles;
}}

\item{probs}{A numeric vector of probabilities with values in \verb{[0, 1]} to use in
combination with \code{cutoff = "quantile"}. By default, it corresponds to quartiles
(\code{c(0, 0.25, 0.5, 0.75, 1)}).}
}
\value{
A tibble with the same dimension as \code{sig_data}, having a column \code{sample_id}
with sample identifiers and one column for each input signature giving
sample classes.
}
\description{
\code{stratify_sig()} is supposed to be used in combination after \code{\link[=hack_sig]{hack_sig()}} in
order to classify your samples in one of two or more signature classes.
}
\examples{
scores <- hack_sig(test_expr, "immune")
stratify_sig(scores)
}
\seealso{
\code{\link[=hack_sig]{hack_sig()}}, \code{\link[stats:quantile]{stats::quantile()}}
}
