% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_expr}
\alias{test_expr}
\title{A toy gene expression matrix}
\format{
A matrix with 20000 random gene symbols as rows and 20 samples as columns.
}
\usage{
test_expr
}
\description{
A gene expression matrix ...
}
\keyword{datasets}
