% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_singscore.R
\name{compute_singscore}
\alias{compute_singscore}
\title{Singscore}
\usage{
compute_singscore()
}
\arguments{
\item{expr_data}{A gene expression matrix (or data frame) with gene symbols as
row names and samples as columns.}

\item{signature}{A list of character vectors corresponding to gene signatures.}
}
\value{
A data frame with
}
\description{
Obtain singscores, as implemented in XXX, from a cohort of samples.
}
\examples{
m <- matrix()
}
