% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ssgsea.R
\name{compute_ssgsea}
\alias{compute_ssgsea}
\title{Single sample GSEA}
\usage{
compute_ssgsea(expr_data, signatures, norm = "raw", alpha = 0.25)
}
\arguments{
\item{expr_data}{A gene expression matrix (or data frame) with gene symbols as
row names and samples as columns.}

\item{signatures}{A list of named character vectors corresponding to gene signatures.}

\item{norm}{Type of normalization affecting the single sample scores. Can be one of:
\itemize{
\item \code{raw}, obtain raw scores (\emph{default});
\item \code{separate}, normalize raw scores across samples for each signature separately.
\item \code{all}, normalize raw scores both across samples and signatures.
}}

\item{alpha}{Exponent in the running sum of the score calculation which weights
the gene ranks. Default to \eqn{\alpha = 0.25}. GSEA classic uses \eqn{\alpha = 1}.}
}
\value{
A tibble with the first column indicating sample identifiers (\code{sample_id})
and all the other columns ....
}
\description{
Obtain single sample GSEA scores for a list of gene signatures,
as implemented in \emph{Barbie et al. (2009)}. This method is a single sample
generalization of the original Gene Set Enrichment Analysis (GSEA) procedure
(see \emph{Subramanian et al. (2005)}).
}
\examples{
set.seed(123)

test_expr <- data.frame(
    sample1 = rnorm(10, mean = 8, sd = 2),
    sample2 = rnorm(10, mean = 7, sd = 2),
    sample3 = rnorm(10, mean = 8, sd = 3),
    row.names = paste0("gene", 1:10)
)

test_list <- list(
    gs1 = paste0("gene", c(1, 2, 6)),
    gs2 = paste0("gene", 1:5),
    gs3 = paste0("gene", c(3, 1, 9)),
    gs4 = paste0("gene", c(1, 3, 9))
)

compute_ssgsea(test_expr, test_list)

}
\references{
Barbie, D. A., Tamayo, P., Boehm, J. S., Kim, S. Y., Moody, S. E., Dunn, I. F.,
Schinzel, A. C., Sandy, P., Meylan, E., Scholl, C., Fröhling, S., Chan, E. M.,
Sos, M. L., Michel, K., Mermel, C., Silver, S. J., Weir, B. A., Reiling, J. H.,
Sheng, Q., Gupta, P. B., … Hahn, W. C. (2009). Systematic RNA interference
reveals that oncogenic KRAS-driven cancers require TBK1. \emph{Nature}, 462(7269),
108–112. \href{https://doi.org/10.1038/nature08460}{doi: 10.1038/nature08460}.

Hänzelmann, S., Castelo, R., & Guinney, J. (2013). GSVA: gene set variation
analysis for microarray and RNA-seq data. \emph{BMC bioinformatics}, 14, 7.
\href{https://doi.org/10.1186/1471-2105-14-7}{doi :10.1186/1471-2105-14-7}.

Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B. L.,
Gillette, M. A., Paulovich, A., Pomeroy, S. L., Golub, T. R., Lander, E. S.,
& Mesirov, J. P. (2005). Gene set enrichment analysis: a knowledge-based
approach for interpreting genome-wide expression profiles.
\emph{Proceedings of the National Academy of Sciences of the United States of America},
102(43), 15545–15550. \href{https://doi.org/10.1073/pnas.0506580102}{doi: 10.1073/pnas.0506580102}.
}
\seealso{

}
