% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_zscore.R
\name{compute_zscore}
\alias{compute_zscore}
\title{Single Sample Z-score}
\usage{
compute_zscore(expr_data, signatures)
}
\arguments{
\item{expr_data}{A gene expression matrix (or data frame) with gene symbols as
row names and samples as columns.}

\item{signature}{A list of character vectors corresponding to gene signatures.}
}
\value{
A data frame with
}
\description{
Obtain single sample z-scores, as implemented in XXX, from a cohort of samples.
}
\examples{
m <- matrix()
}
