% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_zscore.R
\name{compute_zscore}
\alias{compute_zscore}
\title{Single sample z-score}
\usage{
compute_zscore(expr_data, signatures)
}
\arguments{
\item{expr_data}{A gene expression matrix (or data frame) with gene symbols as
row names and samples as columns.}

\item{signatures}{A list of named character vectors corresponding to gene signatures.}
}
\value{
A tibble with the first column indicating sample identifiers (\code{sample_id})
and all the other columns ....
}
\description{
Obtain single sample z-scores for a list of gene signatures,
as implemented in \emph{Lee et al. (2008)}.
}
\examples{
set.seed(123)

test_expr <- data.frame(
    sample1 = rnorm(10, mean = 8, sd = 2),
    sample2 = rnorm(10, mean = 7, sd = 2),
    sample3 = rnorm(10, mean = 8, sd = 3),
    row.names = paste0("gene", 1:10)
)

test_list <- list(
    gs1 = paste0("gene", c(1, 2, 6)),
    gs2 = paste0("gene", 1:5),
    gs3 = paste0("gene", c(3, 1, 9)),
    gs4 = paste0("gene", c(1, 3, 9))
)

compute_zscore(test_expr, test_list)

}
\references{
Hänzelmann, S., Castelo, R., & Guinney, J. (2013). GSVA: gene set variation
analysis for microarray and RNA-seq data. \emph{BMC bioinformatics}, 14, 7.
\href{https://doi.org/10.1186/1471-2105-14-7}{doi :10.1186/1471-2105-14-7}.

Lee, E., Chuang, H. Y., Kim, J. W., Ideker, T., & Lee, D. (2008). Inferring
pathway activity toward precise disease classification.
\emph{PLoS computational biology}, 4(11), e1000217.
\href{https://doi.org/10.1371/journal.pcbi.1000217}{doi: 10.1371/journal.pcbi.1000217}.
}
\seealso{

}
