% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig_info.R
\name{get_sig_info}
\alias{get_sig_info}
\title{Display available gene signatures}
\usage{
get_sig_info()
}
\value{
A tibble with one row per signature and four columns:
\itemize{
\item \code{signature_id}, a unique identifier associated to a signature;
\item \code{signature_keywords}, valid keywords to use in the \code{signatures} argument of
\code{hack_sig()} and \code{check_sig()} as well as in the \code{keywords} argument of
\code{get_sig_genes()};
\item \code{publication_doi}, the original publication DOI;
\item \code{description}, a brief description about the signature.
}
}
\description{
\code{get_sig_info()} returns information about all gene signatures implemented in
\code{hacksig}.
}
\examples{
get_sig_info()
}
\seealso{
\code{\link[=check_sig]{check_sig()}}, \code{\link[=hack_sig]{hack_sig()}}, \code{\link[=get_sig_genes]{get_sig_genes()}}
}
