% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack_cinsarc.R
\name{hack_cinsarc}
\alias{hack_cinsarc}
\title{Classify a cohort of samples based on the CINSARC Signature}
\usage{
hack_cinsarc(expr_data, dm_status)
}
\arguments{
\item{expr_data}{A gene expression matrix (or data frame) with gene symbols as
row names and samples as columns.}

\item{dm_status}{A numeric vector specifying whether a sample has either (1)
or not (0) developed distant metastasis.}
}
\value{
A data frame with
}
\description{
Given a gene expression data frame and a vector indicating the distant metastasis
status of a sample, `hack_cinsarc()` classifies samples in one of two risk
classes, C1 or C2, using the 67-gene signature CINSARC and the LOOCV method
as implemented in ...
}
\examples{
m <- matrix()
}
