% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack_estimate.R
\name{hack_estimate}
\alias{hack_estimate}
\title{ESTIMATE}
\usage{
hack_estimate(expr_data)
}
\arguments{
\item{expr_data}{A gene expression matrix (or data frame) with gene symbols as
row names and samples as columns.}
}
\value{
A data frame with
}
\description{
Obtain Immune, Stroma, ESTIMATE and Purity scores from a cohort of samples.
}
\examples{
m <- matrix()
}
