% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack_estimate.R
\name{hack_estimate}
\alias{hack_estimate}
\title{ESTIMATE}
\usage{
hack_estimate(expr_data)
}
\arguments{
\item{expr_data}{A gene expression matrix (or data frame) with gene symbols as
row names and samples as columns.}
}
\value{
A data frame with
}
\description{
Obtain Immune, Stroma, ESTIMATE and Purity scores from a cohort of samples.
}
\examples{

}
\references{
Yoshihara, K., Shahmoradgoli, M., Martínez, E., Vegesna, R.,
Kim, H., Torres-Garcia, W., Treviño, V., Shen, H., Laird, P. W., Levine, D. A.,
Carter, S. L., Getz, G., Stemke-Hale, K., Mills, G. B., & Verhaak, R. G. (2013).
Inferring tumour purity and stromal and immune cell admixture from
expression data. \emph{Nature communications}, 4, 2612.
\href{https://doi.org/10.1038/ncomms3612}{doi: 10.1038/ncomms3612}
}
