% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack_immunophenoscore.R
\name{hack_immunophenoscore}
\alias{hack_immunophenoscore}
\title{Classify a cohort of samples based on the Immunophenoscore}
\usage{
hack_immunophenoscore(expr_data)
}
\arguments{
\item{expr_data}{A gene expression matrix (or data frame) with gene symbols as
row names and samples as columns.}
}
\value{
A data frame with
}
\description{
WRITE SOMETHING
}
\examples{
m <- matrix()
}
