% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack_immunophenoscore.R
\name{hack_immunophenoscore}
\alias{hack_immunophenoscore}
\title{Hack the Immunophenoscore}
\source{
\href{https://github.com/icbi-lab/Immunophenogram}{github.com/icbi-lab/Immunophenogram}
}
\usage{
hack_immunophenoscore(expr_data, feature = "all")
}
\arguments{
\item{expr_data}{A gene expression matrix (or data frame) with gene symbols as
row names and samples as columns.}

\item{feature}{A string indicating whether you want a more granular output (\code{type})
or a more aggregated one (\code{class}). Each of the two possible choices will give
the raw and discrete immunophenoscores.}
}
\value{
A data frame with
}
\description{
Obtain various immune biomarkers scores, which combined together give the
immunophenoscore (\emph{Charoentong et al., 2017}).
}
\details{
Immunophenoscore (IPS), is a score that was generated, through the use of machine learning, with The Cancer Genome Atlas data
from 20 solid cancers. The Cancer Immunome Atlas was generated (https://tcia.at/) by the depiction of intratumoral immune landscapes.
The score identifies the tumor immunogenicity, and it was able to predict the response to immunocheckpoint inhibitors
(anti-cytotoxic T lymphocyte antigen-4 (CTLA-4) and anti-programmed cell death protein 1 (anti-PD-1) antibodies) in two validation cohorts.
}
\references{
Charoentong, P., Finotello, F., Angelova, M., Mayer, C.,
Efremova, M., Rieder, D., Hackl, H., & Trajanoski, Z. (2017). Pan-cancer
Immunogenomic Analyses Reveal Genotype-Immunophenotype Relationships and
Predictors of Response to Checkpoint Blockade. \emph{Cell reports}, 18(1), 248–262.
\href{https://doi.org/10.1016/j.celrep.2016.12.019}{doi: 10.1016/j.celrep.2016.12.019}.
}
\seealso{
\code{\link[=hack_sig]{hack_sig()}} to compute Immunophenoscore biomarkers in different
ways (use \code{signatures = "ips"}).

\code{\link[=check_sig]{check_sig()}} to check if all/most of the Immunophenoscore biomarkers are
present in your expression matrix (use \code{signatures = "ips"}).
}
