% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hallmark_list}
\alias{hallmark_list}
\title{Hallmark gene set list}
\format{
A named list of 50 signatures corresponding to the Hallmark gene sets.
}
\usage{
hallmark_list
}
\description{
A gene expression matrix ...
}
\keyword{datasets}
