% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signatures_data}
\alias{signatures_data}
\title{Gene-level information for each signature implemented.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{signature_id}{a unique ID given to the signature}
\item{gene_symbol}{Gene Symbol ID}
\item{entrez_gene_id}{Entrez gene ID}
\item{gene_type}{gives the gene group}
\item{gene_weight}{gives the gene weight in the model}
\item{method}{default method implemented in the original work}
}
}
\usage{
signatures_data
}
\description{
A dataset containing information about signatures implemented in the \code{hacksig}
package.
}
\keyword{datasets}
