% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_expr}
\alias{test_expr}
\title{A toy gene expression matrix}
\format{
A random normal data matrix with 20000 genes as rows and 20 samples as columns.
}
\usage{
test_expr
}
\description{
A gene expression matrix simulating expression profiles of 20 samples. It should
be used only for testing purpose.
}
\examples{
class(test_expr)
dim(test_expr)
check_sig(test_expr)
}
\keyword{datasets}
